+++ src/utils.ts
@@ -1,0 +1,27 @@
+import { ServerResponse, IncomingMessage } from 'http';
+
+export const sendResponse = (res: ServerResponse, statusCode: number, data: any) => {
+    res.writeHead(statusCode, { 'Content-Type': 'application/json' });
+    res.end(JSON.stringify(data));
+};
+
+export const parseRequestBody = (req: IncomingMessage): Promise<any> => {
+    return new Promise((resolve, reject) => {
+        let body = '';
+        req.on('data', chunk => {
+            body += chunk.toString();
+        });
+        req.on('end', () => {
+            try {
+                resolve(JSON.parse(body));
+            } catch (err) {
+                reject(err);
+            }
+        });
+    });
+};
+
+export const validateUUID = (id: string): boolean => {
+    const regex = /^[0-9a-f]{8}-[0-9a-f]{4}-[1-5][0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$/i;
+    return regex.test(id);
+};
